+++ Get-RecentBitly.ps1
@@ -1,3 +1,3 @@
 Param($count = 5)
-Get-BitlyGroupLinks -Size $Count
+Get-BitlyGroupLink  -Size $Count
 #| Select-object Created,ID,CustomLinks,Title,Tags,Url