+++ changelog.md
@@ -2,6 +2,18 @@
 
 ## [Unreleased]
 
+## [0.7.0] - 2024-08-07
+
+### Added
+
+- Added command `Export-BitlyConfiguration` [[Issue #3](https://github.com/jdhitsolutions/PSBitly/issues/3)]
+- Added the alias `Export-BitlyToken` to `Save-BitlyToken`.
+- Added missing online help links
+
+### Changed
+
+- Updated `README.md`.
+
 ## [0.6.0] - 2023-10-18
 
 ### Added
@@ -53,5 +65,5 @@
 
 - initial files
 
-[Unreleased]: https://github.com/jdhitsolutions/PSBitly/compare/v0.6.0..HEAD
+[0.7.0]: (https://github.com/jdhitsolutions/PSBitly/compare/v0.6.0..v0.7.0)
 [0.6.0]: https://github.com/jdhitsolutions/PSBitly/compare/v0.5.0..v0.6.0